+++ src/Entities/Schema.php
@@ -16,8 +16,6 @@ use InvalidArgumentException;
 use IteratorAggregate;
 use Rebilly\Entities\Coupons\Coupon;
 use Rebilly\Entities\Coupons\Redemption;
-use Rebilly\Entities\EmailNotifications\EmailNotification;
-use Rebilly\Entities\EmailNotifications\EmailNotificationTracking;
 use Rebilly\Entities\KycDocuments\KycDocument;
 use Rebilly\Entities\Shipping\ShippingZone;
 use Rebilly\Rest\Collection;
@@ -314,18 +312,6 @@ final class Schema implements IteratorAggregate, ArrayAccess
             },
             'tracking/webhooks/{webhookId}' => function (array $content) {
                 return new WebhookTracking($content);
-            },
-            'email-notifications' => function (array $content) {
-                return new Collection(new EmailNotification(), $content);
-            },
-            'email-notifications/{notificationId}' => function (array $content) {
-                return new EmailNotification($content);
-            },
-            'tracking/email-notifications' => function (array $content) {
-                return new Collection(new EmailNotificationTracking(), $content);
-            },
-            'tracking/email-notifications/{trackId}' => function (array $content) {
-                return new EmailNotificationTracking($content);
             },
             'gateway-accounts/{gatewayAccountId}/downtime-schedules' => function (array $content) {
                 return new Collection(new GatewayAccountDowntime(), $content);