+++ src/Entities/Subscription.php
@@ -13,24 +13,7 @@ namespace Rebilly\Entities;
 use Rebilly\Rest\Entity;
 
 /**
- * Class Subscription
- *
- * ```json
- * {
- *   "id"
- *   "customerId"
- *   "planId"
- *   "websiteId"
- *   "initialInvoiceId"
- *   "billingContactId"
- *   "deliveryContactId"
- *   "quantity"
- *   "customFields"
- * }
- * ```
- *
- * @author Veaceslav Medvedev <veaceslav.medvedev@rebilly.com>
- * @version 0.1
+ * Class Subscription.
  */
 final class Subscription extends Entity
 {
@@ -48,18 +31,6 @@ final class Subscription extends Entity
     public function getCustomerId()
     {
         return $this->getAttribute('customerId');
-    }
-
-    /**
-     * @return null|Customer
-     */
-    public function getCustomer()
-    {
-        if ($this->hasEmbeddedResource('customer')) {
-            return new Customer($this->getEmbeddedResource('customer'));
-        } else {
-            return null;
-        }
     }
 
     /**
@@ -81,63 +52,11 @@ final class Subscription extends Entity
     }
 
     /**
-     * @param string $value
-     *
-     * @return Subscription
-     */
-    public function setInitialInvoiceId($value)
-    {
-        return $this->setAttribute('initialInvoiceId', $value);
-    }
-
-    /**
-     * @return string
-     */
-    public function getPlanId()
-    {
-        return $this->getAttribute('planId');
-    }
-
-    /**
-     * @return null|Plan
-     */
-    public function getPlan()
-    {
-        if ($this->hasEmbeddedResource('plan')) {
-            return new Plan($this->getEmbeddedResource('plan'));
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * @param string $value
-     *
-     * @return Subscription
-     */
-    public function setPlanId($value)
-    {
-        return $this->setAttribute('planId', $value);
-    }
-
-    /**
      * @return string
      */
     public function getWebsiteId()
     {
         return $this->getAttribute('websiteId');
-    }
-
-    /**
-     * @return null|Website
-     */
-    public function getWebsite()
-    {
-        if ($this->hasEmbeddedResource('website')) {
-            return new Website($this->getEmbeddedResource('website'));
-        } else {
-            return null;
-        }
     }
 
     /**
@@ -148,92 +67,6 @@ final class Subscription extends Entity
     public function setWebsiteId($value)
     {
         return $this->setAttribute('websiteId', $value);
-    }
-
-    /**
-     * @return int
-     */
-    public function getQuantity()
-    {
-        return $this->getAttribute('quantity');
-    }
-
-    /**
-     * @param int $value
-     *
-     * @return Subscription
-     */
-    public function setQuantity($value)
-    {
-        return $this->setAttribute('quantity', (int) $value);
-    }
-
-    /**
-     * @deprecated The method is deprecated and will be removed in next version.
-     *
-     * @return string
-     */
-    public function getBillingContactId()
-    {
-        return $this->getAttribute('billingContactId');
-    }
-
-    /**
-     * @deprecated The method is deprecated and will be removed in next version.
-     *
-     * @return null|Contact
-     */
-    public function getBillingContact()
-    {
-        if ($this->hasEmbeddedResource('billingContact')) {
-            return new Contact($this->getEmbeddedResource('billingContact'));
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * @deprecated The method is deprecated and will be removed in next version.
-     * @param string $value
-     *
-     * @return $this
-     */
-    public function setBillingContactId($value)
-    {
-        return $this->setAttribute('billingContactId', $value);
-    }
-
-    /**
-     * @return string
-     */
-    public function getDeliveryContactId()
-    {
-        return $this->getAttribute('deliveryContactId');
-    }
-
-    /**
-     * @deprecated The method is deprecated and will be removed in next version.
-     *
-     * @return null|Contact
-     */
-    public function getDeliveryContact()
-    {
-        if ($this->hasEmbeddedResource('deliveryContact')) {
-            return new Contact($this->getEmbeddedResource('deliveryContact'));
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * @deprecated The method is deprecated and will be removed in next version.
-     * @param string $value
-     *
-     * @return $this
-     */
-    public function setDeliveryContactId($value)
-    {
-        return $this->setAttribute('deliveryContactId', $value);
     }
 
     /**
@@ -269,14 +102,6 @@ final class Subscription extends Entity
     }
 
     /**
-     * @return string
-     */
-    public function getCreatedTime()
-    {
-        return $this->getAttribute('createdTime');
-    }
-
-    /**
      * @param string $value
      *
      * @return $this
@@ -284,6 +109,14 @@ final class Subscription extends Entity
     public function setRenewalTime($value)
     {
         return $this->setAttribute('renewalTime', $value);
+    }
+
+    /**
+     * @return string
+     */
+    public function getCreatedTime()
+    {
+        return $this->getAttribute('createdTime');
     }
 
     /**
@@ -363,18 +196,6 @@ final class Subscription extends Entity
     }
 
     /**
-     * @return null|LeadSource
-     */
-    public function getLeadSource()
-    {
-        if ($this->hasEmbeddedResource('leadSource')) {
-            return new LeadSource($this->getEmbeddedResource('leadSource'));
-        } else {
-            return null;
-        }
-    }
-
-    /**
      * @return Address
      */
     public function getBillingAddress()
@@ -390,6 +211,16 @@ final class Subscription extends Entity
     public function setBillingAddress($value)
     {
         return $this->setAttribute('billingAddress', $value);
+    }
+
+    /**
+     * @param array $data
+     *
+     * @return Address
+     */
+    public function createBillingAddress(array $data)
+    {
+        return new Address($data);
     }
 
     /**
@@ -411,6 +242,16 @@ final class Subscription extends Entity
     }
 
     /**
+     * @param array $data
+     *
+     * @return Address
+     */
+    public function createDeliveryAddress(array $data)
+    {
+        return new Address($data);
+    }
+
+    /**
      * @return RiskMetadata
      */
     public function getRiskMetadata()
@@ -429,26 +270,6 @@ final class Subscription extends Entity
     }
 
     /**
-     * @param array $data
-     *
-     * @return Address
-     */
-    public function createBillingAddress(array $data)
-    {
-        return new Address($data);
-    }
-
-    /**
-     * @param array $data
-     *
-     * @return Address
-     */
-    public function createDeliveryAddress(array $data)
-    {
-        return new Address($data);
-    }
-
-    /**
      * @return array
      */
     public function getLineItems()
@@ -462,5 +283,74 @@ final class Subscription extends Entity
     public function getLineItemSubtotal()
     {
         return $this->getAttribute('lineItemSubtotal');
+    }
+
+    /**
+     * @return array|Subscriptions\PlanItem[]
+     */
+    public function getItems()
+    {
+        return $this->getAttribute('items');
+    }
+
+    /**
+     * @param array $value
+     *
+     * @return $this
+     */
+    public function setItems(array $value)
+    {
+        return $this->setAttribute('items', $value);
+    }
+
+    /**
+     * @param array $data
+     *
+     * @return array|Subscriptions\PlanItem[]
+     */
+    public function createItems(array $data)
+    {
+        return array_map(
+            function (array $item) {
+                return new Subscriptions\PlanItem($item);
+            },
+            $data
+        );
+    }
+
+    /**
+     * @return null|Website
+     */
+    public function getWebsite()
+    {
+        $data = $this->getEmbeddedResource('website');
+
+        return $data
+            ? new Website($data)
+            : null;
+    }
+
+    /**
+     * @return null|Customer
+     */
+    public function getCustomer()
+    {
+        $data = $this->getEmbeddedResource('customer');
+
+        return $data
+            ? new Customer($data)
+            : null;
+    }
+
+    /**
+     * @return null|LeadSource
+     */
+    public function getLeadSource()
+    {
+        $data = $this->getEmbeddedResource('leadSource');
+
+        return $data
+            ? new LeadSource($data)
+            : null;
     }
 }